server/requirement.txt
absl-py==0.12.0
apache-beam
astunparse==1.6.3
attrs==21.4.0
avro-python3==1.10.2
cachetools==4.2.4
certifi==2021.10.8
charset-normalizer==2.0.10
colorama
contextlib2==21.6.0
crcmod==1.7
cycler==0.11.0
Cython==0.29.26
dataclasses==0.6
dill==0.3.1.1
dm-tree==0.1.6
docopt==0.6.2
fastavro==1.4.9
flatbuffers==1.12
fonttools==4.28.5
future==0.18.2
gast==0.3.3
gin-config==0.5.0
google-api-core==2.3.2
google-api-python-client==2.34.0
google-auth==2.3.3
google-auth-httplib2==0.1.0
google-auth-oauthlib==0.4.6
google-cloud-bigquery==2.31.0
google-cloud-core==2.2.1
google-crc32c==1.3.0
google-pasta==0.2.0
google-resumable-media==2.1.0
googleapis-common-protos==1.54.0
grpcio==1.43.0
grpcio-status==1.43.0
h5py==2.10.0
hdfs==2.6.0
httplib2==0.19.1
idna==3.3
imgviz==1.4.1
importlib-metadata==4.10.0
importlib-resources==5.4.0
joblib==1.1.0
kaggle==1.5.12
keras==2.7.0
Keras-Preprocessing==1.1.2
kiwisolver==1.3.2
labelme==4.6.0
libclang==12.0.0
lvis==0.5.3
lxml==4.7.1
Markdown==3.3.6
matplotlib==3.2.2
numpy==1.19.5
oauth2client==4.1.3
oauthlib==3.1.1
object-detection
opencv-python==4.5.5.62
opencv-python-headless==4.5.5.62
opt-einsum==3.3.0
orjson==3.6.5
packaging==21.3
pandas==1.3.5
Pillow==9.0.0
portalocker==2.3.2
promise==2.3
proto-plus==1.19.8
protobuf==3.19.2
psutil==5.9.0
py-cpuinfo==8.0.0
pyarrow==6.0.1
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycocotools==2.0.4
pydot==1.4.2
pymongo==3.12.3
pyparsing==2.4.7
PyQt5==5.15.6
PyQt5-Qt5==5.15.2
PyQt5-sip==12.9.0
python-dateutil==2.8.2
python-slugify==5.0.2
pytz==2021.3
PyYAML==6.0
QtPy==2.0.0
regex==2021.11.10
requests==2.27.1
requests-oauthlib==1.3.0
rsa==4.8
sacrebleu==2.0.0
scikit-learn==1.0.2
scipy==1.7.3
sentencepiece==0.1.96
seqeval==1.2.2
six==1.15.0
tabulate==0.8.9
tensorboard==2.7.0
tensorboard-data-server==0.6.1
tensorboard-plugin-wit==1.8.1
tensorflow==2.7.0
tensorflow-addons==0.15.0
tensorflow-datasets==4.4.0
tensorflow-estimator==2.7.0
tensorflow-hub==0.12.0
tensorflow-io==0.23.1
tensorflow-io-gcs-filesystem==0.23.1
tensorflow-metadata==1.5.0
tensorflow-model-optimization==0.7.0
tensorflow-text==2.7.3
termcolor==1.1.0
text-unidecode==1.3
tf-models-official==2.4.0
tf-slim==1.1.0
threadpoolctl==3.0.0
tqdm==4.62.3
typeguard==2.13.3
typing-extensions==3.7.4.3
uritemplate==4.1.1
urllib3==1.26.8
Werkzeug==2.0.2
wrapt==1.12.1
zipp==3.7.0
